services/model_runtime/requirements.txt
numpy==2.0.2
Pillow==11.1.0
opencv-python-headless==4.11.0.86
scipy==1.13.1
PyYAML==6.0.2
